型変更したら更新できなくなったのでuseApi使用し、機能をフックにまとまる
