+++ src/app/_types/apiRequests/dashboard/setting/updateRequest.ts
@@ -2,15 +2,12 @@ import { Gender } from "@prisma/client";
 import { ApiRequest } from "../../apiRequest";
 
 /**現在の月齢を表示*/
-export interface Baby {
+export interface UpdateRequests {
   name: string;
   birthday: Date;
   expectedDateOfBirth: Date;
   birthWeight: number;
   gender: Gender;
-}
-export interface UpdateRequests extends ApiRequest {
-  data: Baby;
 }
 
 /**赤ちゃん初回登録時に赤ちゃんIDを登録するためUserテーブル更新 */
